Refactoring logback enactment logger
