+++ src/main/java/at/uibk/dps/ee/enactables/logging/logback/LogbackEnactmentLogger.java
@@ -1,5 +1,7 @@
-package at.uibk.dps.ee.enactables.logging;
+package at.uibk.dps.ee.enactables.logging.logback;
 
+import at.uibk.dps.ee.enactables.logging.EnactmentLogEntry;
+import at.uibk.dps.ee.enactables.logging.EnactmentLogger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -7,6 +9,9 @@ import org.slf4j.LoggerFactory;
  * The {@link LogbackEnactmentLogger} is used to log information about the enactment. Depending
  * on the configuration of logback this data may be logged to STDOUT, a log file or a
  * relational database.
+ *
+ * @author Markus Moosbrugger
+ *
  */
 public class LogbackEnactmentLogger implements EnactmentLogger {
   protected final Logger logger = LoggerFactory.getLogger(LogbackEnactmentLogger.class);
